cleanup of db schema
