+++ prisma/schema.prisma
@@ -304,6 +304,7 @@ model NetworkHandler {
   FeeReceiverNetworkWallet AdminWallet   @relation("FeeReceiverNetworkWallet", fields: [adminWalletId], references: [id], onDelete: Cascade)
   adminWalletId            String
   feePermille              Int           @default(50)
+  cooldownTime             Int           @default(600000)
 
   PaymentRequests  PaymentRequest[]  @relation("PaymentRequestHandler")
   PurchaseRequests PurchaseRequest[] @relation("PurchaseRequestNetworkHandler")
