+++ src/routes/api/purchases/index.ts
@@ -77,6 +77,7 @@ export const queryPurchaseRequestGet = payAuthenticatedEndpointFactory.build({
 });
 
 export const createPurchaseInitSchemaInput = z.object({
+    id: z.string(),
     blockchainIdentifier: z.string().max(250).describe("The identifier of the purchase. Is provided by the seller"),
     network: z.nativeEnum($Enums.Network).describe("The network the transaction will be made on"),
     sellerVkey: z.string().max(250).describe("The verification key of the seller"),
@@ -135,6 +136,7 @@ export const createPurchaseInitPost = payAuthenticatedEndpointFactory.build({
 
 
 export const refundPurchaseSchemaInput = z.object({
+    id: z.string(),
     blockchainIdentifier: z.string().max(250).describe("The identifier of the purchase to be refunded"),
     network: z.nativeEnum($Enums.Network).describe("The network the Cardano wallet will be used on"),
     paymentContractAddress: z.string().max(250).optional().describe("The address of the smart contract holding the purchase"),
